.github/requirements.txt
PyGithub==2.1.1
openai==1.51.0
google-generativeai==0.8.3
anthropic==0.37.1
PyYAML==6.0.1
requests==2.31.0
httpx==0.28.1
